Changed MovementScript to PlayerScript. Started working in Wizbowo.
PlayerScript makes more sense since it is no longer just a movement script. Wizbowo's animations are set to just be a random either wand wave or slam attack, the issue is that niether is actually doing anything other than moving the sprite. I am simply trying to get the projectiles, or "WizbowoBullets" to shoot out when the animation starts but that isn't working rn.